requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
einops==0.8.0
executing==2.1.0
fastapi==0.115.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
HyperPyYAML==1.2.2
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.8
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langsmith==0.1.131
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.51.0
openai-whisper @ git+https://github.com/openai/whisper.git@25639fc17ddc013d56c594bfbf7644f2185fad84
optuna==4.0.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytube==15.0.0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.4.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.4.1
torchmetrics==1.4.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
widgetsnbextension==4.0.13
yarl==1.13.1
yt-dlp==2024.9.27